move LTT summoners to companion object
